# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Api covid19 Documentation",
        "description": "Implementation of Swagger with in Laravel",
        "contact": {
            "email": "admin@admin.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/homintec/penalite": {
            "get": {
                "tags": [
                    "Tests"
                ],
                "summary": "Get List Of penlaite",
                "description": "Returns all penlaite",
                "operationId": "index",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "not found"
                    }
                }
            },
            "post": {
                "tags": [
                    "store"
                ],
                "summary": "save penalite",
                "description": "Login by email, password",
                "operationId": "store",
                "requestBody": {
                    "description": "pass penalite",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "rouland",
                                    "type",
                                    "facteur",
                                    "penalite",
                                    "autorise",
                                    "excedant"
                                ],
                                "properties": {
                                    "rouland": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "string"
                                    },
                                    "type": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "string"
                                    },
                                    "facteur": {
                                        "type": "string",
                                        "example": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "422": {
                        "description": "Wrong credentials response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Sorry, wrong email address or password. Please try again"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    }
}